requirements.txt
pip==21.2.4
pyTelegramBotAPI==4.0.1
python-decouple==3.5
requests==2.5.0
setuptools==49.2.1
